Handle passing null arguments
* Added unit tests for those code parts

+++ lib/private/subadmin.php
@@ -65,7 +65,11 @@ class SubAdmin extends PublicEmitter {
 	 * @param IGroup $group group $user becomes subadmin of
 	 * @return bool
 	 */
-	public function createSubAdmin(IUser $user, IGroup $group) {
+	public function createSubAdmin(IUser $user = null, IGroup $group = null) {
+		if ($user === null || $group === null) {
+			return false;
+		}
+
 		$qb = $this->dbConn->getQueryBuilder();
 
 		$result = $qb->insert('group_admin')
@@ -86,7 +90,11 @@ class SubAdmin extends PublicEmitter {
 	 * @param IGroup $group the group
 	 * @return bool
 	 */
-	public function deleteSubAdmin(IUser $user, IGroup $group) {
+	public function deleteSubAdmin(IUser $user = null, IGroup $group = null) {
+		if ($user === null || $group === null) {
+			return false;
+		}
+
 		$qb = $this->dbConn->getQueryBuilder();
 
 		$result = $qb->delete('group_admin')
@@ -104,7 +112,11 @@ class SubAdmin extends PublicEmitter {
 	 * @param IUser $user the SubAdmin
 	 * @return IGroup[]
 	 */
-	public function getSubAdminsGroups(IUser $user) {
+	public function getSubAdminsGroups(IUser $user = null) {
+		if ($user === null) {
+			return [];
+		}
+
 		$qb = $this->dbConn->getQueryBuilder();
 
 		$result = $qb->select('gid')
@@ -125,7 +137,11 @@ class SubAdmin extends PublicEmitter {
 	 * @param IGroup $group the group
 	 * @return IUser[]
 	 */
-	public function getGroupsSubAdmins(IGroup $group) {
+	public function getGroupsSubAdmins(IGroup $group = null) {
+		if ($group === null) {
+			return [];
+		}
+
 		$qb = $this->dbConn->getQueryBuilder();
 
 		$result = $qb->select('uid')
@@ -168,7 +184,11 @@ class SubAdmin extends PublicEmitter {
 	 * @param IGroup $group
 	 * @return bool
 	 */
-	public function isSubAdminofGroup(IUser $user, IGroup $group) {
+	public function isSubAdminofGroup(IUser $user = null, IGroup $group = null) {
+		if ($user === null || $group === null) {
+			return false;
+		}
+
 		$qb = $this->dbConn->getQueryBuilder();
 
 		$result = $qb->select('*')
@@ -185,10 +205,11 @@ class SubAdmin extends PublicEmitter {
 	 * @param IUser $user 
 	 * @return bool
 	 */
-	public function isSubAdmin($user) {
+	public function isSubAdmin(\OCP\IUser $user = null) {
 		if (is_null($user)) {
 			return false;
 		}
+
 		// Check if the user is already an admin
 		if ($this->groupManager->isAdmin($user->getUID())) {
 			return true;
@@ -212,7 +233,11 @@ class SubAdmin extends PublicEmitter {
 	 * @param IUser $user
 	 * @return bool
 	 */
-	public function isUserAccessible($subadmin, $user) {
+	public function isUserAccessible($subadmin = null, $user = null) {
+		if ($subadmin === null || $user === null) {
+			return false;
+		}
+
 		if(!$this->isSubAdmin($subadmin)) {
 			return false;
 		}
